Refactored code to seperate files based on function
