+++ model.py
@@ -1,110 +1,7 @@
 import torch
-import os
-
 import torch.nn as nn
-import torch.optim as optim
-import numpy as np
-import nibabel as nib
-
-from torchvision import transforms
-from tqdm import tqdm
-
-
-def to_channels(arr: np.ndarray, dtype=np.uint8) -> np.ndarray:
-    channels = np.unique(arr)
-    res = np.zeros(arr.shape + (len(channels),), dtype=dtype)
-    for c in channels:
-        c = int(c)
-        res[..., c : c + 1][arr == c] = 1
-
-    return res
-
-
-def load_data_3D(
-    imageNames,
-    normImage=False,
-    categorical=False,
-    dtype=np.float32,
-    getAffines=False,
-    orient=False,
-    early_stop=False,
-):
-    """
-    Load medical image data from names , cases list provided into a list for each .
-    This function pre - allocates 5D arrays for conv3d to avoid excessive memory &
-    usage .
-    normImage : bool ( normalise the image 0.0 -1.0)
-    orient : Apply orientation and resample image ? Good for images with large slice &
-    thickness or anisotropic resolution
-    dtype : Type of the data . If dtype =np.uint8 , it is assumed that the data is &
-    labels
-    1early_stop : Stop loading pre - maturely ? Leaves arrays mostly empty , for quick &
-    loading and testing scripts .
-    """
-    affines = []
-
-    # ~ interp = ' continuous '
-    interp = "linear "
-    if dtype == np.uint8:  # assume labels
-        interp = "nearest "
-
-    # get fixed size
-    num = len(imageNames)
-    niftiImage = nib.load(imageNames[0])
-    if orient:
-        niftiImage = im.applyOrientation(niftiImage, interpolation=interp, scale=1)
-        # ~ testResultName = "oriented.nii.gz"
-        # ~ niftiImage.to_filename(testResultName)
-    first_case = niftiImage.get_fdata(caching="unchanged")
-    if len(first_case.shape) == 4:
-        first_case = first_case[:, :, :, 0]  # sometimes extra dims, remove
-    if categorical:
-        first_case = utils.to_channels(first_case, dtype=dtype)
-        rows, cols, depth, channels = first_case.shape
-        images = np.zeros((num, rows, cols, depth, channels), dtype=dtype)
-    else:
-        rows, cols, depth = first_case.shape
-        images = np.zeros((num, rows, cols, depth), dtype=dtype)
-
-    for i, inName in enumerate(tqdm(imageNames)):
-        niftiImage = nib.load(inName)
-        if orient:
-            niftiImage = im.applyOrientation(niftiImage, interpolation=interp, scale=1)
-        inImage = niftiImage.get_fdata(caching="unchanged")  # read disk only
-        affine = niftiImage.affine
-        if len(inImage.shape) == 4:
-            inImage = inImage[:, :, :, 0]  # sometimes extra dims in HipMRI_study data
-        inImage = inImage[:, :, :depth]  # clip slices
-        inImage = inImage.astype(dtype)
-        if normImage:
-            # ~ inImage = inImage / np.linalg.norm(inImage)
-            # ~ inImage = 255. * inImage / inImage.max()
-            inImage = (inImage - inImage.mean()) / inImage.std()
-        if categorical:
-            inImage = utils.to_channels(inImage, dtype=dtype)
-            # ~ images[i, :, :, :, :] = inImage
-            images[
-                i,
-                : inImage.shape[0],
-                : inImage.shape[1],
-                : inImage.shape[2],
-                : inImage.shape[3],
-            ] = inImage  # with pad
-        else:
-            # ~ images[i, :, :, :] = inImage
-            images[i, : inImage.shape[0], : inImage.shape[1], : inImage.shape[2]] = (
-                inImage  # with pad
-            )
-
-        affines.append(affine)
-        if i > 20 and early_stop:
-            break
-
-    if getAffines:
-        return images, affines
-    else:
-        return images
-
+# import numpy as np
+# from torchvision import transforms
 
 class NormConv3D(nn.Module):
     def __init__(
@@ -258,21 +155,6 @@ class UNet3D(nn.Module):
         return out
     
 
-# Define device
-device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
-if not torch.cuda.is_available():
-    print("Warning: No access to CUDA, model being trained on CPU.")
 
-model = UNet3D.to(device)
-
-# Define the loss function and optimizer
-criterion = nn.CrossEntropyLoss()
-optimizer = optim.Adam(model.parameters(), lr=1e-4) # TODO learning scheduler?
-
-slice_data_path = "/home/groups/comp3710/HipMRI_Study_open/keras_slices_data"
-semantic_labels_path = "/home/groups/comp3710/HipMRI_Study_open/semantic_labels_only"
-semantic_MR_path = "/home/groups/comp3710/HipMRI_Study_open/semantic_MRs"
-
-# TODO load in dataS
